+++ src/Shader.hpp
@@ -26,8 +26,8 @@ class Shader {
         unsigned int getRendererID() const { return rendererID; }
 
         // Set uniforms, TODO: use templates to have multiple types of uniforms
-        void setUniform1i(const std::string& name, float value);
-        void setUniform1f(const std::string& name, int value);
+        void setUniform1i(const std::string& name, int value);
+        void setUniform1f(const std::string& name, float value);
         void setUniform3f(const std::string& name, float v0, float v1, float v2);
         void setUniform4f(const std::string& name, float v0, float v1, float v2, float v3);
         void setUniformVec3(const std::string& name, const glm::vec3& vec);
